fix(tf,keras): int32/int64 mask dtype in MAY features; robust epsilon-trainable assertion
The MAY maclaurin_features mask compared an int32 range against int64 degrees,
which TensorFlow rejects (`cannot compute Less ... int32 vs int64`). Cast both
sides to int32 in the tf and keras backends.

Also harden the keras lazy training-step test: epsilon_param initializes at
inverse-softplus(1e-5) ≈ -11.5, a saturated point where the softplus gradient is
~1e-5, so its value barely moves in two SGD steps. Assert epsilon receives a
real (non-None, non-zero) gradient — the property #37 requires (trainable, not
frozen) — instead of a fragile value-change threshold.

Verified locally with TensorFlow 2.21 / Keras 3.14 (py3.12): test_tf 129 passed,
test_keras 130 passed.

+++ src/nmn/keras/performer_yat.py
@@ -201,8 +201,9 @@ def maclaurin_features(
     proj = ops.einsum("...d,mld->...ml", x, omegas)
 
     # mask[m, l] = l < N_m  →  (M, nmax); neutralize unused factors with 1.0.
-    l = ops.arange(nmax)
-    mask = ops.expand_dims(l, 0) < ops.expand_dims(degrees, 1)   # (M, nmax)
+    # Cast both sides to int32 so the comparison never mixes int32/int64.
+    l = ops.cast(ops.arange(nmax), "int32")
+    mask = ops.expand_dims(l, 0) < ops.cast(ops.expand_dims(degrees, 1), "int32")
     mask = ops.cast(mask, "bool")
     proj = ops.where(mask, proj, ops.ones_like(proj))
